fix(kernel): keep the DSV4 indexer logits row offset in int64
The portable MXFP4 scorer writes a [tokens, max_candidates] FP32 workspace
and formed the row offset as an int32 product of the program id and the
candidate bound. A decode batch of a few thousand rows scheduled next to a
million-token request (262K compressed candidates) pushes that workspace
past 2**31 elements, and Triton does not promote the int32 product, so the
store would wrap. Extend the row index to int64 before the multiply; the
page-table and cache offsets already were.

The overflow regime needs an 8 GB workspace and cannot be exercised by the
simulator tests; the existing numerical coverage confirms the store path is
unchanged for the shapes it can run.

+++ tokenspeed-kernel/python/tokenspeed_kernel/ops/attention/dsv4/triton/indexer.py
@@ -119,7 +119,7 @@ def _mxfp4_logits_kernel(
     weights = tl.load(weights_ptr + token * HEADS + heads)
     scores = tl.sum(tl.maximum(products, 0.0) * weights[:, None], axis=0)
     tl.store(
-        logits_ptr + token * max_candidates + offsets,
+        logits_ptr + token.to(tl.int64) * max_candidates + offsets,
         tl.where(valid, scores, -float("inf")),
         mask=offsets < max_candidates,
     )
